refactor: move currencies update logic to the hook
