+++ src/App.tsx
@@ -1,9 +1,6 @@
-import React, { ReactNode, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
+import React, { ReactNode } from 'react';
 
-import { getCurrencies } from '@/services/currency/getCurrencies';
-import { updateCurrencies } from '@/store/slices/currencySlice';
-import { CurrencyRated } from '@/types/currency';
+import { useAutomaticCurrenciesUpdate } from '@/hooks/useAutomaticCurrenciesUpdate';
 
 import '@/styles/index.scss';
 
@@ -12,25 +9,10 @@ interface AppProps {
 }
 
 export const App = ({ children }: AppProps) => {
-  const dispatch = useDispatch();
+  const intervalTimeMinutes = 20;
+  const intervalTimeMs = intervalTimeMinutes * 60 * 1000;
 
-  useEffect(() => {
-    const currencies = localStorage.getItem('persist:root');
-    if (!currencies) {
-      getCurrencies().then((data: CurrencyRated[]) => {
-        dispatch(updateCurrencies(data || []));
-      });
-    }
-
-    const intervalTimeMs = 20 * 60 * 1000;
-    const intervalId = setInterval(() => {
-      getCurrencies().then((data: CurrencyRated[]) => {
-        dispatch(updateCurrencies(data || []));
-      });
-    }, intervalTimeMs);
-
-    return () => clearInterval(intervalId);
-  }, []);
+  useAutomaticCurrenciesUpdate(intervalTimeMs);
 
   return <div className='app'>{children}</div>;
 };
