+++ src/constants/colorTheme/colorTheme.ts
@@ -1,0 +1,5 @@
+export const colorThemeKey = 'color-theme';
+export const colorThemes = {
+  dark: 'dark',
+  light: 'light',
+};
