+++ src/hooks/useTheme.ts
@@ -1,18 +1,23 @@
 import { useLayoutEffect, useState } from 'react';
 
+import { colorThemeKey, colorThemes } from '@/constants/colorTheme/colorTheme';
+
 export const useTheme = () => {
   const [theme, setTheme] = useState(
-    localStorage.getItem('color-theme') || 'dark',
+    localStorage.getItem(colorThemeKey) || colorThemes.dark,
   );
 
   const toggleTheme = (isChecked: boolean) => {
-    const newTheme = isChecked ? 'dark' : 'light';
+    const newTheme = isChecked ? colorThemes.dark : colorThemes.light;
     setTheme(newTheme);
-    localStorage.setItem('color-theme', newTheme);
+    localStorage.setItem(colorThemeKey, newTheme);
   };
 
   useLayoutEffect(() => {
-    document.documentElement.setAttribute('color-theme', theme || 'dark');
+    document.documentElement.setAttribute(
+      colorThemeKey,
+      theme || colorThemeKey,
+    );
   }, [theme]);
 
   return { theme, toggleTheme };
